Remove all the cout calls used before to control the integration
Integration is finally working, but the step must be small if the diffusion constant is high, so far only equilibrium between A (area where the motors can't bind to MTs), and B (area where motors can bind to Mts). Now I will add the term of binding in the equation and see, after then include koff and the movement on the Mts.