Things are moving!
Included movement of motors also in the differential equations. Steady state works well, but there is an underestimation in the beginning caused by the fact that treadmilling of the motors is not yet a steady state, so at early time points we have more motors bound to the minus end "than there should be", a better approximation is needed to describe the kinetics of the system.