Output Input Branch created
So far the numerical deterministic calculation is not accurate, it does not include the k_off properly. I will work on that soon, but first I want to include a method to use output and input files. Also low diffusion regimes prevent the prediction to be accurate, maybe I could even include the kinetics of the depletion of the areas close to the MTs. Probably it is going to be a while until I can run the simulation life while it also creates the config file, but we will see!