Input file working roughly
So far it is working, starting to implement new classes that are useful, like the properties of each of the type of objects in a different class from the object itself. This is useful and relates to the structure of the config file in cytosim (first you use set -> define some properties of a type of object, and give a name to this particular set of objects of a certain class, for instance in class ball, you write set ball motor1 to create certain properties that will be transmitted to the balls you create using new X motor1) . Francois suggested including a hierarchy, for objects and for properties. There is a superclass (Object, Property), there are subclasses tube, ball, properties of the tube. This will allow us to use a function that can operate in any object, for instance to obtain the position, or to do other operations. I guess I will see the potential of this method as I continue developing the simulation.

The use of pointers is still a bit confusing to me in some cases, I should work on it to understand it better.