Parser finally working, Object oriented structure implemented
The parser is working, but requires extensive cleaning from superfluous lines that make no sense at the moment. It would also need new features in the future if we want to include more complex instructions that cannot be described by prop = value. We will see.

Right now the simulation is very similar to how it was before, but somehow the binding doesn't work, I will revise that tomorrow.