Input finished, output coming soon
Things starting to look ok, however there are still some things to curate in the input parser, for instance until now I was taking length of the MTs as a parameter, but actually I should use it as a different property, since it should change if we eventually incorporate the growth of fibres. This is relevant because my parser would not look for { } in the "new" statements, and would require necessarily an { } for the "set" statements. This should be changed. Also there is still a lot of lines that don't do anything in the input.h file, this will be addressed once everything is working. Probably at that point I will realise that I am doing similar things in different ways at different parts of the program, this should be harmonised also.

As for the output file, I am working at the same time in the functions that generate the output file and in the ones that read it, this way new inconveniences keep emerging, but things are moving forward, and I have an idea of how I want to do things.