Almost there
The parser has been updated, now seems better. Population of the simulation seems to be working. The only problem now is that the program seems to stop when I call scanf() on some of the lines. Why does this happen? I really don't understand why, but maybe tomorrow I will find the solution. Once I manage to make the scanf work, it should be easier to store all the info of the simulation somehow, and then access it, rather than reading it from the file every time.