Done!
Finally, it is possible to store the information in a folder, and read it from there. It still is a dirty solution, and a lot of things should be fixed (I am using global variables for things that should be done with a "run" statement, for instance. Also still you cannot choose which color your motors or Mts will have), and that is only regarding the config file.

The output file is not very good yet, for instance now the objects properties are not written on the objects, since for the display they are not important (they will when you can choose the colors of the fibres or motors). Also, in every step there is information that is read from the file that is superfluous, and anyway it would be better to store in the memory all the information of the simulation, and access it there rather than read from the file all the time, I would say.

Also, the output file should be a "read only" file, since one would want that by accident one cannot change a value in it, I will have to see how to do this from C++ directly.

Finally I would need to make a program in Python to plot the information, I would have to see how to do that.