BaSim_2
This is actually the correct solution, before I wasn't doing the 8 vs 1/8 thing. This is done because in the previous solution, whether a ball would bind or not didn't depend on how many Mts were close to it, however in this case we do an approach that overcomes this problem and allows for checking binding first, and then checking available Mts.