Vectorizing the coordinates and their operations
Still some left to change, also I want to include the movement along the MT by the balls as a reference to its minus end, so that the Mts can move, and make it possible to balls to move on them as they move without mayor complicated calculations. However when they start to shrink and grow from both sides, its going to be hard!